# Conflicts:
#	Ubihard_Project/Bin/AnimationLibrary/Animation.h
#	Ubihard_Project/Bin/AnimationLibrary/Skeleton.h
#	Ubihard_Project/Bin/FBXLoader/x64/Debug/FBXLoader.tlog/FBXLoader.write.1u.tlog
#	Ubihard_Project/Bin/FBXLoader/x64/Debug/vc140.idb
#	Ubihard_Project/Bin/x64/Debug/FBXLoader.exp
#	Ubihard_Project/Ubihard_Project/x64/Debug/Main.obj
#	Ubihard_Project/Ubihard_Project/x64/Debug/Scene.obj
#	Ubihard_Project/Ubihard_Project/x64/Debug/Ubihard_Project.log
#	Ubihard_Project/Ubihard_Project/x64/Debug/Ubihard_Project.tlog/CL.write.1.tlog
#	Ubihard_Project/Ubihard_Project/x64/Debug/Ubihard_Project.tlog/Ubihard_Project.write.1u.tlog
#	Ubihard_Project/Ubihard_Project/x64/Debug/Ubihard_Project.tlog/link.read.1.tlog
#	Ubihard_Project/Ubihard_Project/x64/Debug/vc140.idb
#	Ubihard_Project/Ubihard_Project/x64/Debug/vc140.pdb

+++ Ubihard_Project/Ubihard_Project/Model.cpp
@@ -230,8 +230,8 @@ void Model::LightCulling()
 	ID3D11ShaderResourceView* nullSRV = NULL;
 	ID3D11DepthStencilView* nullDSV = NULL;
 	const float res = 16.0f;
-	const unsigned int NUM_TILES_X = (float)CLIENT_WIDTH / res;
-	const unsigned int NUM_TILES_Y = (float)CLIENT_HEIGHT / res;
+	const unsigned int NUM_TILES_X = (unsigned int)((float)CLIENT_WIDTH / res);
+	const unsigned int NUM_TILES_Y = (unsigned int)((float)CLIENT_HEIGHT / res);
 
 	//set render target and dsv to null
 	devContext->OMSetRenderTargets(1, &nullRTV, nullDSV);
